Cleanup remaining moved comment references
