+++ library/Vanilla/Quill/Blots/TextBlot.php
@@ -7,11 +7,7 @@
 
 namespace Vanilla\Quill\Blots;
 
-use Vanilla\Quill\Blots\Embeds\EmojiBlot;
-use Vanilla\Quill\Formats;
 use Vanilla\Quill\BlotGroup;
-use Vanilla\Quill\FormattableTextTrait;
-use Vanilla\Quill\Renderer;
 
 class TextBlot extends AbstractBlot {
